+++ crates/web-client/src/export.rs
@@ -69,4 +69,15 @@ impl WebClient {
             Err(JsValue::from_str("Client not initialized"))
         }
     }
+
+    /// Retrieves the entire underlying web store and returns it as a JsValue
+    ///
+    /// Meant to be used in conjunction with the force_import_store method
+    pub async fn export_store(&mut self) -> Result<JsValue, JsValue> {
+        let store = self.store.as_ref().ok_or(JsValue::from_str("Store not initialized"))?;
+        let export =
+            store.export_store().await.map_err(|err| JsValue::from_str(&format!("{err}")))?;
+
+        Ok(export)
+    }
 }