+++ crates/web-client/test/webClientTestUtils.ts
@@ -342,6 +342,7 @@ export const consumeTransaction = async (
 interface SetupWalletFaucetResult {
   accountId: string;
   faucetId: string;
+  accountHash: string;
 }
 
 export const setupWalletAndFaucet =
@@ -363,6 +364,7 @@ export const setupWalletAndFaucet =
 
       return {
         accountId: account.id().to_string(),
+        accountHash: account.hash().to_hex(),
         faucetId: faucetAccount.id().to_string(),
       };
     });